Revised the code according to the review.
