+++ automata/base/animation.py
@@ -1,8 +1,12 @@
+from collections.abc import Iterable
 from dataclasses import dataclass
-from typing import Iterable, Optional
+from functools import partial
+from typing import Optional, TypeVar
 
 import manim
 import pygraphviz as pgv
+
+_POINTS_IN_INCH = 72
 
 
 class Animate:
@@ -10,7 +14,11 @@ class Animate:
 
     DEFAULT_COLOR = manim.WHITE
     HIGHLIGHT_COLOR = manim.RED
-    manim.Circle.set_default(color=DEFAULT_COLOR)
+    M = TypeVar("M", bound=manim.Mobject)
+
+    @classmethod
+    def default_init(cls, mobject_class: type[M]) -> type[M]:
+        return partial(mobject_class, color=cls.DEFAULT_COLOR)
 
     @classmethod
     def to_default_color(cls, mobject: manim.VMobject) -> manim.ApplyMethod:
@@ -47,11 +55,9 @@ class Animate:
 @dataclass(eq=False)
 class _ManimNode(manim.VGroup):
     """
-    The `ManimNode` class represents a state in FA as a node in `FAGraph` by representing
-    the `Node` object in the `AGraph` object generated by `fa.show_diagram()` with
-    manim.
+    The `ManimNode` class represents a `Node` object in the `AGraph` object with manim.
 
-    The `ManimNode` class is a `VGroup`. For the convenience, it's also added with
+    The `ManimNode` class is a `VGroup`. For the convenience, it's also added with:
 
     Parameters
     ----------
@@ -66,7 +72,7 @@ class _ManimNode(manim.VGroup):
     shape: manim.Dot | manim.Circle | manim.VGroup
     label: Optional[manim.Text]
 
-    def __init__(self, node: pgv.Node):
+    def __init__(self, node: pgv.Node) -> None:
         """
         Parameters
         ----------
@@ -84,22 +90,26 @@ class _ManimNode(manim.VGroup):
         super().__init__(name=node.name)
         radius = float(node.attr["height"]) / 2
         if node.attr["shape"] == "point":
-            self.shape = manim.Dot(radius=radius)
+            self.shape = Animate.default_init(manim.Dot)(radius=radius)
             self.add(self.shape)
         elif node.attr["shape"].endswith("circle"):
-            circle = manim.Circle(radius=radius)
+            circle = Animate.default_init(manim.Circle)(radius=radius)
             self.shape = (
                 manim.VGroup(
                     circle,
-                    manim.Circle().surround(circle, buffer_factor=0.8),
+                    Animate.default_init(manim.Circle)().surround(
+                        circle, buffer_factor=0.8
+                    ),
                 )
                 if node.attr["shape"].startswith("double")
                 else circle
             )
             self.add(self.shape)
-            self.label = manim.Text(node.name, font_size=float(node.attr["fontsize"]))
+            self.label = Animate.default_init(manim.Text)(
+                node.name, font_size=float(node.attr["fontsize"])
+            )
             self.add(self.label)
-        x, y = (float(pt) / 72 for pt in node.attr["pos"].split(","))
+        x, y = (float(pt) / _POINTS_IN_INCH for pt in node.attr["pos"].split(","))
         self.set_x(x)
         self.set_y(y)
 
@@ -107,11 +117,9 @@ class _ManimNode(manim.VGroup):
 @dataclass(eq=False)
 class _ManimEdge(manim.VGroup):
     """
-    The `ManimEdge` class represents a transition in FA as an edge in `FAGraph` by
-    representing the `Edge` object in the `AGraph` object generated by
-    `fa.show_diagram()` with manim.
+    The `ManimEdge` class represents an `Edge` object in the `AGraph` object with manim.
 
-    The `ManimEdge` class is a `VGroup`. For the convenience, it's also added with
+    The `ManimEdge` class is a `VGroup`. For the convenience, it's also added with:
 
     Parameters
     ----------
@@ -125,7 +133,7 @@ class _ManimEdge(manim.VGroup):
     edge: manim.VGroup
     label: Optional[manim.Text]
 
-    def __init__(self, edge: pgv.Edge):
+    def __init__(self, edge: pgv.Edge) -> None:
         r"""
         Parameters
         ----------
@@ -142,8 +150,10 @@ class _ManimEdge(manim.VGroup):
         self.edge = self.__parse_spline(edge.attr["pos"].replace("\\\r", ""))
         self.add(self.edge)
         if label_text := edge.attr["label"]:
-            self.label = manim.Text(label_text, font_size=float(edge.attr["fontsize"]))
-            x, y = (float(pt) / 72 for pt in edge.attr["lp"].split(","))
+            self.label = Animate.default_init(manim.Text)(
+                label_text, font_size=float(edge.attr["fontsize"])
+            )
+            x, y = (float(pt) / _POINTS_IN_INCH for pt in edge.attr["lp"].split(","))
             self.label.set_x(x)
             self.label.set_y(y)
             self.add(self.label)
@@ -164,14 +174,17 @@ class _ManimEdge(manim.VGroup):
         """
         points = edge_pos.split()
         control_points = tuple(
-            (*(float(pt) / 72 for pt in point.split(",")), 0) for point in points[1:]
+            (*(float(pt) / _POINTS_IN_INCH for pt in point.split(",")), 0)
+            for point in points[1:]
         )
         result = manim.VGroup()
         for i in range(0, len(control_points) - 1, 3):
-            result.add(manim.CubicBezier(*control_points[i : i + 4]))
-        endp = (*(float(pt) / 72 for pt in points[0].split(",")[1:]), 0)
+            result.add(
+                Animate.default_init(manim.CubicBezier)(*control_points[i : i + 4])
+            )
+        endp = (*(float(pt) / _POINTS_IN_INCH for pt in points[0].split(",")[1:]), 0)
         result.add(
-            manim.Arrow(
+            Animate.default_init(manim.Arrow)(
                 start=control_points[-1],
                 end=endp,
                 max_tip_length_to_length_ratio=1,
@@ -185,7 +198,7 @@ class _ManimInput(manim.VGroup):
     string are separated with each character is generated to a `Text` object, so that
     you can get each character simply with `[]` operator."""
 
-    def __init__(self, text: str):
+    def __init__(self, text: str) -> None:
         """
         Generate the input symbols and put them on the top left of the screen.
 
@@ -247,7 +260,9 @@ class _ManimInput(manim.VGroup):
         Write
             The animation of writing the result for the `Scene` object to `play`.
         """
-        result = manim.Text(f"→ {'accept' if accept else 'reject'}")
+        result = Animate.default_init(manim.Text)(
+            f"→ {'accept' if accept else 'reject'}"
+        )
         if len(self):
             result.next_to(self)
         else:
