+++ automata/fa/fa.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import abc
 import os
 from collections import defaultdict
-from typing import Any, Dict, Generator, Iterable, List, Optional, Set, Tuple, Union
+from typing import Any, Dict, Generator, List, Optional, Set, Tuple, Union
 
 from automata.base.automaton import Automaton, AutomatonStateT
 from automata.base.utils import (
@@ -18,14 +18,12 @@ from automata.base.utils import (
 # Optional imports for use with visual functionality
 try:
     import coloraide
-    import manim
     import pygraphviz as pgv
-
-    from automata.base.animation import Animate, _ManimEdge, _ManimNode
 except ImportError:
     _visual_imports = False
 else:
     _visual_imports = True
+
 
 FAStateT = AutomatonStateT
 
@@ -217,146 +215,3 @@ class FA(Automaton, metaclass=abc.ABCMeta):
         state_set.add(new_state)
 
         return new_state
-
-
-if _visual_imports:
-
-    class _FAGraph(manim.VDict):
-        """
-        The `FAGraph` class is the common class for FA to draw a picture of a `FA`
-        object by representing the `AGraph` object generated by `fa.show_diagram()` with
-        manim.
-
-        The `FAGraph` class is a `VDict`. The keys are the states (with type `FAStateT`)
-        and the transitions (with type `tuple[FAStateT, FAStateT]`, which is their start
-        and end points). Specially, there is another `None` key corresponding to the
-        `nullnode` in the `FA.show_diagram`, which is the start point of the arrow
-        pointing to the initial state. The values are the Manim objects.
-        """
-
-        def __init__(self, fa: FA):
-            """
-            Generate the graph and put it on the down side o the screen.
-
-            Parameters
-            ----------
-            fa : FA
-                The FA object based to create the graph.
-
-                Note that states are discriminated by `FA._get_state_name`, so states
-                are not allowed to have same names. Plus, states are not allowed to be
-                named 'None'.
-            """
-            name_state = {FA._get_state_name(state): state for state in fa.states}
-            fa_diagram = fa.show_diagram()
-            super().__init__(
-                {
-                    name_state.get(node): _ManimNode(node)
-                    for node in fa_diagram.nodes_iter()
-                }
-                | {
-                    tuple(map(name_state.get, edge)): _ManimEdge(edge)
-                    for edge in fa_diagram.edges_iter()
-                }
-            )
-            self.scale_to_fit_width(manim.config.frame_width)
-            self.align_on_border(manim.LEFT, buff=0).align_on_border(manim.DOWN)
-
-        def highlight_states(
-            self, states: Iterable[FAStateT]
-        ) -> Iterable[manim.ApplyMethod]:
-            """
-            Parameters
-            ----------
-            states : Iterable[FAStateT]
-                the states to highlight.
-            Returns
-            -------
-            `Iterable[ApplyMethod]`
-                The animations for the `Scene` object to `play`.
-            """
-            yield from (
-                Animate.highlight(self[new_state])
-                for new_state in states
-                if new_state is not None
-            )
-
-        def change_states(
-            self, old_states: Iterable[FAStateT], new_states: Iterable[FAStateT]
-        ) -> Iterable[manim.ApplyMethod]:
-            """
-            Turn `old_states` to default color and highlight the `new_states`.
-
-            States occured in both `old_states` and `new_states` will still be
-            highlighted.
-
-            Parameters
-            ----------
-            old_states : Iterable[FAStateT]
-                The states to turn to default color.
-            new_states : Iterable[FAStateT]
-                The states to highlight.
-
-            Returns
-            -------
-            Iterable[ApplyMethod]
-                The animations for the `Scene` object to `play`.
-            """
-            yield from (
-                Animate.to_default_color(self[old_state]) for old_state in old_states
-            )
-            yield from self.highlight_states(new_states)
-
-        def change_transitions(
-            self,
-            old_transitions: Iterable[tuple[FAStateT, FAStateT]],
-            new_transitions: Iterable[tuple[FAStateT, FAStateT]],
-        ) -> Iterable[manim.ApplyMethod]:
-            """
-            Turn `old_transitions` to default color and highlight the `new_transitions`.
-
-            Transitions occured in both `old_transitions` and `new_transitions` will
-            still be highlighted.
-
-            Parameters
-            ----------
-            old_transitions : Iterable[FAStateT]
-                The transitions to turn to default color.
-            new_transitions : Iterable[FAStateT]
-                The transitions to highlight.
-
-            Returns
-            -------
-            Iterable[ApplyMethod]
-                The animations for the `Scene` object to `play`.
-            """
-            yield from (
-                Animate.to_default_color(self[old_transition])
-                for old_transition in old_transitions
-            )
-            yield from (
-                Animate.highlight(self[new_transition])
-                for new_transition in new_transitions
-                if new_transition[-1] is not None
-            )
-
-        def clean(
-            self, transitions: Iterable[tuple[FAStateT, FAStateT]]
-        ) -> Iterable[manim.ApplyMethod]:
-            """
-            Cancel all the highlighted elements.
-
-            Parameters
-            ----------
-            transitions : Iterable[tuple[FAStateT, FAStateT]]
-                The highlighed transitions to cancel the highlight.
-
-            Returns
-            -------
-            The animations for `Scene` object to `play`.
-            """
-            yield from (
-                Animate.to_default_color(self[transition])
-                for transition in transitions
-                if transition[1] is not None
-            )
